agents/orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.6",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "openai>=1.108.1",
    "openai-agents>=0.3.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.0.2",
    "starlette>=0.48.0",
    "uvicorn>=0.36.0"
]


[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]

